# maize/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "API documentation for Product endpoints",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/historical-yields": {
            "get": {
                "tags": [
                    "HistoricalYields"
                ],
                "summary": "Get list of all historical yields",
                "description": "Display a listing of the historical yields.",
                "operationId": "90fe2c313c3ed6256ae0960729e7526e",
                "responses": {
                    "200": {
                        "description": "List of historical yields"
                    }
                }
            },
            "post": {
                "tags": [
                    "HistoricalYields"
                ],
                "summary": "Create a new historical yield",
                "description": "Store a newly created historical yield in storage.",
                "operationId": "963a525ed76563a7ed0a3a4929f555e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "region_or_field",
                                    "year",
                                    "yield_t_ha",
                                    "source"
                                ],
                                "properties": {
                                    "region_or_field": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "yield_t_ha": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "source": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Historical yield created"
                    }
                }
            }
        },
        "/historical-yields/{historical_yield}": {
            "get": {
                "tags": [
                    "HistoricalYields"
                ],
                "summary": "Get a specific historical yield",
                "description": "Display the specified historical yield.",
                "operationId": "93d90743eb7e1534a5611d6ab55d59be",
                "parameters": [
                    {
                        "name": "historical_yield",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historical yield details"
                    },
                    "404": {
                        "description": "Historical yield not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "HistoricalYields"
                ],
                "summary": "Update a historical yield",
                "description": "Update the specified historical yield in storage.",
                "operationId": "6cb7c23e47a0eea50b4f6b26aee59f48",
                "parameters": [
                    {
                        "name": "historical_yield",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "region_or_field": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "yield_t_ha": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "source": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Historical yield updated"
                    },
                    "404": {
                        "description": "Historical yield not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "HistoricalYields"
                ],
                "summary": "Delete a historical yield",
                "description": "Remove the specified historical yield from storage.",
                "operationId": "e96373e099a696ec8cdc67c236f047e0",
                "parameters": [
                    {
                        "name": "historical_yield",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Historical yield deleted"
                    },
                    "404": {
                        "description": "Historical yield not found"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get list of products",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "List of products"
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    },
                                    "firmware_version": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created"
                    }
                }
            }
        },
        "/products/{product}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get a product by ID",
                "operationId": "960e17d05ed63211b20fb4b8505f1263",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product data"
                    }
                }
            },
            "put": {
                "tags": [
                    "products"
                ],
                "summary": "Update a product",
                "operationId": "d8f8f68cbee2f6a5c291d7940a3dc4ed",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    },
                                    "firmware_version": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "operationId": "13d9c94a21e9746d2a61b6a3247b7693",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}